pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nx_yaml"
version = "0.0.5"
description = "Fast Representation Graph for PyYAML using NetworkX"
dependencies = [
  "pyyaml>=6",
]

[project.urls]
"Source" = "https://github.com/yaml-programming/nx_yaml"
